get connection

+++ BCCCBackend/BCCCBackend/Prop.cs
@@ -9,10 +9,10 @@ namespace BCCCBackend
     class Prop
     {
         public string Name { get; set; }
-        public double Rate { get; set; }
+        public decimal Rate { get; set; }
         public string Date { get; set; }
         public string Time { get; set; }
-        public double Ask { get; set; }
-        public double Bid { get; set; }
+        public decimal Ask { get; set; }
+        public decimal Bid { get; set; }
     }
 }
